requirements.txt
flask
Flask-SQLAlchemy
uuid
pillow